TriangularReshapes

Collection of functions to efficiently reshape vectors into (strictly) lower triangular matrices and vice versa.

Functionality

TriangularReshapes.lower_triang_to_vector! — Function

lower_triang_to_vector!(v::AbstractVector{T}, M::AbstractMatrix{T}) where {T}

Overwrites the first n * (n + 1) / 2 elements of v with the values of the lower triangular part of M.

Arguments

v: The vector to be overwritten.

M: The matrix whose lower triangular part is used to overwrite the first n * (n + 1) / 2 elements of v.

lower_triang_to_vector!(v, v1, v2, c1, c2, adding)

Overwrites the first n * (n + 1) / 2 elements of v with the values of the lower triangular part of v1*transpose(v2).

Arguments

v: The vector to be overwritten.

v1, v2: Vectors to form the rank-one-matrix v1*v2'

c1, c2: Booleans to indicate whether the vectors v1 or v2 should be conjugated.

adding: Boolean to indicate whether the output should be added to v or if v should be overwritten.

TriangularReshapes.vector_to_lower_triang! — Function

vector_to_lower_triang!(M::AbstractMatrix{T}, v::AbstractVector{T}) where {T}

Overwrites the lower triangular part of M with the values of v.

Arguments

M: The matrix to be overwritten.

v: The vector to be used to overwrite the lower triangular part of M.

TriangularReshapes.lower_triang_to_vector — Function

lower_triang_to_vector(M::AbstractMatrix{T}) where {T}

Returns a vector containing the values of the lower triangular part of M.

Arguments

M: The matrix whose lower triangular part is used to create the vector.

Outputs

v: The vector containing the values of the lower triangular part of M.

TriangularReshapes.vector_to_lower_triang — Function

vector_to_lower_triang(v::AbstractVector)

Creates a matrix M, whose lower triangular part is filled with the values of v.

Arguments

v: The vector to be used to overwrite the lower triangular part of M.

Outputs

M: The matrix whose lower triangular part is filled with the values of v.

TriangularReshapes.vector_to_strictly_lower_triang! — Function

vector_to_strictly_lower_triang!(M::AbstractMatrix{T}, v::AbstractVector{T}) where {T}

Overwrites the strictly lower triangular part of M with the values of v.

Arguments

M: The matrix to be overwritten.

v: The vector to be used to overwrite the strictly lower triangular part of M.

TriangularReshapes.strictly_lower_triang_to_vector! — Function

strictly_lower_triang_to_vector!(v::AbstractVector{T}, M::AbstractMatrix{T}) where {T}

Overwrites the first n * (n - 1) / 2 elements of v with the values of the strictly lower triangular part of M.

Arguments

v: The vector to be overwritten.

M: The matrix whose strictly lower triangular part is used to overwrite the first n * (n - 1) / 2 elements of v.

strictly_lower_triang_to_vector!(v, v1, v2, c1, c2, adding)

Overwrites the first n * (n - 1) / 2 elements of v with the values of the strictly lower triangular part of v1*transpose(v2).

Arguments

v: The vector to be overwritten.

v1, v2: Vectors to form the rank-one-matrix v1*v2'

c1, c2: Booleans to indicate whether the vectors v1 or v2 should be conjugated.

adding: Boolean to indicate whether the output should be added to v or if v should be overwritten.

subtracting: Boolean to indicate whether the output should be subtracted from v or if v should be overwritten. Note that adding and subtracting cannot be both true.

TriangularReshapes.strictly_lower_triang_to_vector — Function

strictly_lower_triang_to_vector(M::AbstractMatrix{T}) where {T}

Returns a vector containing the values of the strictly lower triangular part of M.

Arguments

M: The matrix whose strictly lower triangular part is used to create the vector.

Outputs

v: The vector containing the values of the strictly lower triangular part of M.

TriangularReshapes.vector_to_strictly_lower_triang — Function

vector_to_strictly_lower_triang(v::AbstractVector)

Creates a matrix M, whose strictly lower triangular part is filled with the values of v.

Arguments

v: The vector to be used to overwrite the strictly lower triangular part of M.

Outputs

M: The matrix whose strictly lower triangular part is filled with the values of v.